seeding: update use katello's content management
This commit changes the management of the content from using
a foreman puppet environment to defining and publishing the
content using a katello product, repository and content view.
It will then associate a content source, view and lifecycle
environment with the hostgroup that will be used by provisioning.

For now, the name for the objects is hard coded (e.g. Fusor, Puppet);
however, we could alter it in the future, if needed, to allow it to
be specified by the installer or configuration.  Also, the content will
only be published to the library.