seeding: updates based on fusor-server updates for hostgroups
When a user performs a deployment using the fusor server, they
will provide information such as organization, lifecycle environment,
product to be deployed...etc.  Since those characteristics will
impact the content used by the deployment as specified in the hostgroup,
we have decided that the fusor-server itself will handle the creation
and updating of hostgroups that specify the content details
(e.g. lifecycle environment, content view, operating system,
installation media...etc).  As a result, we no longer need to seed
those hostgroups with the installer; therefore, this commit is to
remove that and align with fusor-server.