installer wizard: add support for configuring BMC feature
If a discovered host is capable of BMC, the host created will include
the interface for it.  If the server is not configured for the BMC feature,
it will cause the provisioning of the host to fail with an error.

While the user can enable the BMC feature after the server is installed,
this change will allow them to enable it during the initial installation.